migrated matrix subspaces related test cases to test_subspaces.py
